test(db-conformance): extracted storage integration tests into a parameterised adapter conformance suite
