+++ packages/db-postgres/vitest.config.ts
@@ -3,8 +3,13 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig(({ mode }) => {
   // Unit:        vitest run --mode=node          → *.test.node.ts (currently none)
   // Integration: vitest run --mode=integration   → **/*.test.ts under src/**/tests/
+  //              plus tests/**/*.test.ts — the conformance entry
+  //              (tests/conformance.integration.test.ts) that runs the
+  //              shared @byline/db-conformance storage suite.
   const isIntegration = mode === 'integration'
-  const testFiles = isIntegration ? ['src/**/tests/**/*.test.ts'] : ['**/*.test.node.ts']
+  const testFiles = isIntegration
+    ? ['src/**/tests/**/*.test.ts', 'tests/**/*.test.ts']
+    : ['**/*.test.node.ts']
 
   return {
     test: {
